Move Scala syntax into pipeline.syntax package
